Simplify the public API by removing traits.
Instead of defining traits whose only purpose was to be mixed in
the relevant `autoImport` objects, we define their members directly
inside the `autoImport`s.

This highlighted that the recommended compatibility imports were

  import sbtcrossproject.{crossProject, CrossType}

whereas in fact they should always have been

  import sbtcrossproject.CrossPlugin.autoImport.{crossProject, CrossType}

instead. This commit also adapts the readme accordingly. Deprecated
members in the package object maintain source compatibility.
