+++ sbt-crossproject/src/main/scala/sbtcrossproject/CrossPlugin.scala
@@ -1,5 +1,7 @@
 package sbtcrossproject
 
+import scala.language.experimental.macros
+
 import sbt._
 import sbt.KeyRanks.BSetting
 
@@ -13,9 +15,20 @@ object CrossPlugin extends AutoPlugin {
   @deprecated("Use autoImport instead.", "0.5.0")
   val AutoImport = autoImport
 
-  object autoImport extends JVMCross with CrossProjectExtra {
+  object autoImport {
 
+    type CrossType = sbtcrossproject.CrossType
     val CrossType = sbtcrossproject.CrossType
+
+    // The crossProject macro
+
+    @deprecated("use crossProject(JSPlatform, JVMPlatform)", "0.1.0") def crossProject: CrossProject.Builder =
+      macro CrossProjectMacros.oldCrossProject_impl
+
+    def crossProject(platforms: Platform*): CrossProject.Builder =
+      macro CrossProjectMacros.vargCrossProject_impl
+
+    // Cross-classpath dependency builders
 
     final implicit def toCrossClasspathDependencyConstructor(
         cp: CrossProject): CrossClasspathDependency.Constructor =
@@ -24,5 +37,24 @@ object CrossPlugin extends AutoPlugin {
     final implicit def toCrossClasspathDependency(
         cp: CrossProject): CrossClasspathDependency =
       new CrossClasspathDependency(cp, None)
+
+    // The JVM platform
+
+    val JVMPlatform = sbtcrossproject.JVMPlatform
+
+    implicit def JVMCrossProjectBuilderOps(
+        builder: CrossProject.Builder): JVMCrossProjectOps =
+      new JVMCrossProjectOps(builder)
+
+    implicit class JVMCrossProjectOps(project: CrossProject) {
+      def jvm: Project = project.projects(JVMPlatform)
+
+      def jvmSettings(ss: Def.SettingsDefinition*): CrossProject =
+        jvmConfigure(_.settings(ss: _*))
+
+      def jvmConfigure(transformer: Project => Project): CrossProject =
+        project.configurePlatform(JVMPlatform)(transformer)
+    }
+
   }
 }
