+++ sbt-crossproject/src/main/scala/sbtcrossproject/CrossProjectMacros.scala
@@ -2,10 +2,9 @@ package sbtcrossproject
 
 import java.io.File
 import scala.reflect.macros.Context
-import scala.language.experimental.macros
 
-object CrossProjectExtra {
-  private[sbtcrossproject] def crossProject_impl(c: Context)(
+private[sbtcrossproject] object CrossProjectMacros {
+  def crossProject_impl(c: Context)(
       platformsArgs: List[c.Expr[Platform]]): c.Expr[CrossProject.Builder] = {
     import c.universe._
 
@@ -78,15 +77,4 @@ object CrossProjectExtra {
     import c.universe._
     crossProject_impl(c)(platforms.toList)
   }
-}
-
-trait CrossProjectExtra {
-  @deprecated("use crossProject(JSPlatform, JVMPlatform)", "0.1.0") def crossProject: CrossProject.Builder =
-    macro CrossProjectExtra.oldCrossProject_impl
-
-  def crossProject(platforms: Platform*): CrossProject.Builder =
-    macro CrossProjectExtra.vargCrossProject_impl
-
-  private[sbtcrossproject] def nonEmpty(s: String, label: String): Unit =
-    require(s.trim.length > 0, label + " cannot be empty.")
 }
