--- sbt-scalajs-crossproject/src/main/scala/scalajscrossproject/ScalaJSCross.scala
@@ -1,27 +1,0 @@
-package scalajscrossproject
-
-import sbtcrossproject._
-
-import sbt._
-
-import org.scalajs.sbtplugin.ScalaJSPlugin
-
-import scala.language.implicitConversions
-
-trait ScalaJSCross {
-  val JSPlatform = scalajscrossproject.JSPlatform
-
-  implicit def JSCrossProjectBuilderOps(
-      builder: CrossProject.Builder): JSCrossProjectOps =
-    new JSCrossProjectOps(builder.crossType(CrossType.Full))
-
-  implicit class JSCrossProjectOps(project: CrossProject) {
-    def js: Project = project.projects(JSPlatform)
-
-    def jsSettings(ss: Def.SettingsDefinition*): CrossProject =
-      jsConfigure(_.settings(ss: _*))
-
-    def jsConfigure(transformer: Project => Project): CrossProject =
-      project.configurePlatform(JSPlatform)(transformer)
-  }
-}
