+++ sbt-scalajs-crossproject/src/main/scala/scalajscrossproject/ScalaJSCrossPlugin.scala
@@ -1,23 +1,33 @@
 package scalajscrossproject
 
-import org.scalajs.sbtplugin.{ScalaJSCrossVersion, ScalaJSPlugin}
-import org.scalajs.sbtplugin.impl.ScalaJSGroupID
-
-import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
+import scala.language.implicitConversions
 
 import sbt._
+import sbtcrossproject._
 
-import scala.language.implicitConversions
+import org.scalajs.sbtplugin.ScalaJSPlugin
 
 object ScalaJSCrossPlugin extends AutoPlugin {
   override def trigger           = allRequirements
   override def requires: Plugins = ScalaJSPlugin
 
-  object autoImport extends ScalaJSCross
+  object autoImport {
 
-  import autoImport._
+    val JSPlatform = scalajscrossproject.JSPlatform
 
-  override def projectSettings: Seq[Setting[_]] = Seq(
-    platformDepsCrossVersion := ScalaJSCrossVersion.binary
-  )
+    implicit def JSCrossProjectBuilderOps(
+        builder: CrossProject.Builder): JSCrossProjectOps =
+      new JSCrossProjectOps(builder.crossType(CrossType.Full))
+
+    implicit class JSCrossProjectOps(project: CrossProject) {
+      def js: Project = project.projects(JSPlatform)
+
+      def jsSettings(ss: Def.SettingsDefinition*): CrossProject =
+        jsConfigure(_.settings(ss: _*))
+
+      def jsConfigure(transformer: Project => Project): CrossProject =
+        project.configurePlatform(JSPlatform)(transformer)
+    }
+
+  }
 }
